GitHub_Road_Freight_Risk_AI_Starter/requirements.txt
streamlit==1.35.0
streamlit-folium==0.16.0
streamlit-js-eval==0.1.5
folium==0.15.1
requests==2.31.0
pandas==2.2.2
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
